Switch to English meal names in table formatting
The method of extracting meal names was changed in the table formatting and api schema files. Functionality of these scripts remain the same, but instead of using the German meal name (`name`), the English version (`name_en`) is now used. This change makes the application more accessible for English speaking users.